Struct zmq_mio::Socket

pub struct Socket { /* fields omitted */ }

Asynchronous ØMQ socket.

Methods

impl Socket

pub fn new(socket: Socket) -> Self

Create a new event-wrapped ØMQ socket. Takes an existing zmq::Socket instance as an only argument.

pub fn as_raw_fd(&self) -> Result<RawFd>

Returns an io::Result with the raw socket file-descriptor.

pub fn get_ref(&self) -> &Socket

Returns a reference to the underlying zmq::Socket. Useful for setting socket options at runtime.

pub fn bind(&self, address: &str) -> Result<()>

Bind the socket to the given address.

pub fn connect(&self, address: &str) -> Result<()>

Connect the socket to the given address.

pub fn set_subscribe(&self, prefix: &[u8]) -> Result<()>

Subscribe this socket to the given prefix.

pub fn send<T>(&self, item: T, flags: i32) -> Result<()> where
T: Sendable,

Send a message.

Due to the provided From implementations, this works for &[u8], Vec<u8> and &str, as well as zmq::Message itself.

Any flags set will be combined with zmq::DONTWAIT, which is needed for non-blocking mode. The internal zmq::Error::EAGAIN is automatically translated to io::ErrorKind::WouldBlock, which you MUST handle without failing.

pub fn send_multipart<I, T>(&self, iter: I, flags: i32) -> Result<()> where
I: IntoIterator<Item = T>,
T: Into<Message>,

Send a multi-part message. Takes any iterator of valid message types.

Due to the provided From implementations, this works for &[u8], Vec<u8> and &str, as well as zmq::Message itself.

Any flags set will be combined with zmq::DONTWAIT, which is needed for non-blocking mode. The internal zmq::Error::EAGAIN is automatically translated to io::ErrorKind::WouldBlock, which you MUST handle without failing.

pub fn get_rcvmore(&self) -> Result<bool>

Return true if there are more frames of a multipart message to receive.

pub fn recv(&self, msg: &mut Message, flags: i32) -> Result<()>

Read a single zmq::Message from the socket. Any flags set will be combined with zmq::DONTWAIT, which is needed for non-blocking mode. The internal zmq::Error::EAGAIN is automatically translated to io::ErrorKind::WouldBlock, which you MUST handle without failing.

Any flags set will be combined with zmq::DONTWAIT, which is needed for non-blocking mode. The internal zmq::Error::EAGAIN is automatically translated to io::ErrorKind::WouldBlock, which you MUST handle without failing.

pub fn recv_into(&self, msg: &mut [u8], flags: i32) -> Result<usize>

Receive bytes into a slice. The length passed to zmq_recv is the length of the slice. The return value is the number of bytes in the message, which may be larger than the length of the slice, indicating truncation.

Any flags set will be combined with zmq::DONTWAIT, which is needed for non-blocking mode. The internal zmq::Error::EAGAIN is automatically translated to io::ErrorKind::WouldBlock, which you MUST handle without failing.

pub fn recv_msg(&self, flags: i32) -> Result<Message>

Receive a message into a fresh zmq::Message.

Any flags set will be combined with zmq::DONTWAIT, which is needed for non-blocking mode. The internal zmq::Error::EAGAIN is automatically translated to io::ErrorKind::WouldBlock, which you MUST handle without failing.

pub fn recv_bytes(&self, flags: i32) -> Result<Vec<u8>>

Receive a message as a byte vector.

Any flags set will be combined with zmq::DONTWAIT, which is needed for non-blocking mode. The internal zmq::Error::EAGAIN is automatically translated to io::ErrorKind::WouldBlock, which you MUST handle without failing.

pub fn recv_string(&self, flags: i32) -> Result<Result<String, Vec<u8>>>

Receive a String from the socket.

If the received message is not valid UTF-8, it is returned as the original Vec in the Err part of the inner result.

Any flags set will be combined with zmq::DONTWAIT, which is needed for non-blocking mode. The internal zmq::Error::EAGAIN is automatically translated to io::ErrorKind::WouldBlock, which you MUST handle without failing.

pub fn recv_multipart(&self, flags: i32) -> Result<Vec<Vec<u8>>>

Receive a multipart message from the socket.

Note that this will allocate a new vector for each message part; for many applications it will be possible to process the different parts sequentially and reuse allocations that way.

Any flags set will be combined with zmq::DONTWAIT, which is needed for non-blocking mode. The internal zmq::Error::EAGAIN is automatically translated to io::ErrorKind::WouldBlock, which you MUST handle without failing.

pub fn get_socket_type(&self) -> Result<SocketType>

Get the SocketType